feat(mobile): split auth into a login screen and a separate sign-up page
Reworks the front door around a reference concept: the login screen leads with
the fast ways back into an account that exists — Apple/Google, a new top-level
"Continue with phone" button, then email — with "Create account" as the line to
the other door. The sign-up page is that other door: the same ways in pointed at
registering, plus the guest option, and "I have an account" back to login.

The guest way in moves off the login screen and onto the sign-up page (ADR-006
addendum). It stays early and never behind a form — one tap behind "Create
account" — but a returning member no longer sees it on the first screen.

Both screens are one shared `AuthFlow({ flow })`: the OTP and password forms are
identical, and two copies of a login form is how one drifts from the other. The
old in-form sign-in/sign-up toggle is gone — the two doors are the choice now.
Routing: sign-up is a public route that a session bounces off like sign-in.
